# backend/docs/swagger.yaml
basePath: /
definitions:
  docs.ErrorResponse:
    properties:
      error:
        example: Message d'erreur
        type: string
    type: object
...
  dtos.CategoryResponse:
    description: Informations catégorie
    properties:
      createdAt:
        description: Date de création
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        description: UUID de la catégorie
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        description: Nom de la catégorie
        example: Visage
        type: string
      updatedAt:
        description: Date de mise à jour
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
...
  dtos.PatchProductRequest:
    description: Permet de mettre à jour uniquement certains champs d'un produit (tous
      les champs sont optionnels)
    properties:
      categoryID:
        description: ID de la catégorie (optionnel, peut être null pour supprimer
          la catégorie)
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      description:
        description: Description du produit (optionnel)
        example: Crème hydratante pour peau sensible
        type: string
      imageURL:
        description: URL de l'image du produit (optionnel)
        example: https://example.com/image.jpg
        type: string
      name:
        description: Nom du produit (optionnel)
        example: Crème hydratante
        type: string
      price:
        description: Prix en euros (optionnel, doit être > 0 si fourni)
        example: 29.99
        type: number
      stock:
        description: Quantité en stock (optionnel, doit être >= 0 si fourni)
        example: 50
        type: integer
    type: object
  dtos.ProductRequest:
    description: Informations produit pour création/modification
    properties:
      categoryID:
        description: ID de la catégorie (optionnel)
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      description:
        description: Description du produit
        example: Crème hydratante pour peau sensible
        type: string
      imageURL:
        description: URL de l'image du produit
        example: https://example.com/image.jpg
        type: string
      name:
        description: Nom du produit
        example: Crème hydratante
        type: string
      price:
        description: Prix en euros (doit être > 0)
        example: 29.99
        type: number
      stock:
        description: Quantité en stock
        example: 50
        minimum: 0
        type: integer
    required:
    - name
    - price
    type: object
  dtos.ProductResponse:
    description: Informations produit avec catégorie
    properties:
      category:
        allOf:
        - $ref: '#/definitions/dtos.CategoryResponse'
        description: Informations de la catégorie (si disponible)
      categoryID:
        description: ID de la catégorie
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      createdAt:
        description: Date de création
        example: "2024-01-01T00:00:00Z"
        type: string
      description:
        description: Description
        example: Crème hydratante pour peau sensible
        type: string
      id:
        description: UUID du produit
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      imageURL:
        description: URL de l'image
        example: https://example.com/image.jpg
        type: string
      name:
        description: Nom du produit
        example: Crème hydratante
        type: string
      price:
        description: Prix en euros
        example: 29.99
        type: number
      stock:
        description: Quantité en stock
        example: 50
        type: integer
      updatedAt:
        description: Date de mise à jour
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: "API REST complète pour une boutique de produits de soins (skincare
    shop) avec authentification JWT et gestion des rôles.\nCette API permet de gérer
    les produits, catégories, commandes et utilisateurs.\n\n## Authentification\nLa
    plupart des endpoints nécessitent une authentification JWT. Pour tester les endpoints
    protégés :\n\n**Étape 1 : Obtenir un token**\n1. Utilisez `/auth/register` pour
    créer un compte OU `/auth/login` pour vous connecter\n2. Copiez le token JWT retourné
    dans la réponse (champ \"token\")\n\n**Étape 2 : Autoriser les requêtes dans Swagger
    UI**\n1. Cliquez sur le bouton **\"Authorize\"** (icône de cadenas) en haut à
    droite de cette page\n2. Dans le champ \"Value\", entrez : `Bearer <votre-token>`
    (remplacez `<votre-token>` par le token copié)\n3. Cliquez sur **\"Authorize\"**
    puis **\"Close\"**\n\n**Étape 3 : Tester les endpoints**\nTous les endpoints protégés
    sont maintenant accessibles. Vous pouvez cliquer sur \"Try it out\" et tester
    chaque endpoint.\n\n## Rôles\n- **USER** : Peut voir les produits, créer des commandes,
    voir ses propres commandes\n- **ADMIN** : Accès complet à toutes les ressources\n\n---\n\n##
    \U0001F510 Compte de test Admin\n\n**Email:** `momo@ynov.com`\n**Password:** `Password2025`\n\nUtilisez
    ce compte pour tester toutes les fonctionnalités admin dans l'API."
  title: PORELO Skincare Shop API
...
paths:
...
  /admin/products:
    post:
      consumes:
      - application/json
      description: Crée un nouveau produit (admin uniquement)
      parameters:
      - description: Informations du produit
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.ProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "403":
          description: Accès refusé - Admin requis
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "409":
          description: Produit avec ce nom existe déjà
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Créer un produit
      tags:
      - Products
  /admin/products/{id}:
    delete:
      consumes:
      - application/json
      description: Supprime un produit (admin uniquement)
      parameters:
      - description: ID du produit
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "403":
          description: Accès refusé - Admin requis
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Supprimer un produit
      tags:
      - Products
    patch:
      consumes:
      - application/json
      description: 'Met à jour uniquement les champs fournis d''un produit (admin
        uniquement). Exemple: changer uniquement l''image sans modifier les autres
        champs.'
      parameters:
      - description: ID du produit
        in: path
        name: id
        required: true
        type: string
      - description: Champs à mettre à jour (tous optionnels)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.PatchProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "403":
          description: Accès refusé - Admin requis
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mettre à jour partiellement un produit
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Met à jour un produit existant (admin uniquement)
      parameters:
      - description: ID du produit
        in: path
        name: id
        required: true
        type: string
      - description: Nouvelles informations du produit
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.ProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "403":
          description: Accès refusé - Admin requis
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mettre à jour un produit
      tags:
      - Products
...
  /products:
    get:
      consumes:
      - application/json
      description: Récupère la liste de tous les produits disponibles avec leurs catégories
        (authentification requise)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.ProductResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Liste tous les produits
      tags:
      - Products
  /products/{id}:
    get:
      consumes:
      - application/json
      description: Récupère les détails d'un produit par son ID (authentification
        requise)
      parameters:
      - description: ID du produit
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Détails d'un produit
      tags:
      - Products
...
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" suivi d''un espace et du token JWT. Exemple: "Bearer
      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"